address PR notes
