+++ packages/api/src/api/users/index.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { post } from './post';
-import { addUser } from '../../utils/addUser';
+import { addUser } from '../../middleware/userMiddleware';
 
 export const users = Router();
