+++ packages/api/tests/api/users.ts/index.test.ts
@@ -9,10 +9,26 @@ jest.mock('../../../src/api/users/post', () => ({
 
 describe('/users router registrations', () => {
   describe('post requests', () => {
+    it('calls next when a valid session is found', async () => {
+      await jest.isolateModulesAsync(async () => {
+        const { users } = require('../../../src/api/users');
+
+        const router = express();
+
+        router.use((req, res, next) => {
+          req.session = { email: 'pancakes@waffles.bananas' } as any;
+          next();
+        }, users);
+
+        const res = await supertest(router).post('/');
+        expect(res.statusCode).toBe(200);
+      });
+    });
+
     it('returns a 401 when a valid session cannot be found', async () => {
       await jest.isolateModulesAsync(async () => {
-        const { users } = await import('../../../src/api/users');
-        
+        const { users } = require('../../../src/api/users');
+
         const router = express();
         router.use(users);
