+++ packages/api/tests/middleware/userMiddleware.test.ts
@@ -1,12 +1,12 @@
 import express, { Request, Response } from 'express';
 import supertest from 'supertest';
-import { addUser } from '../../src/utils/addUser';
+import { addUser } from '../../src/middleware/userMiddleware';
+
 jest.mock('../../src/api/users/post', () => ({
   post: (req: Request, res: Response) => {
     res.sendStatus(200);
   },
 }));
-
 
 const mockUser = {
   firstName: 'a',
@@ -30,27 +30,23 @@ const mockReq:any = {
 const mockRes = {
   sendStatus: jest.fn()
 };
-const mockNext = jest.fn()
+const mockNext = jest.fn();
 
 describe('mounting user on /user', () => {
   it('calls next when a valid session is found', async () => {
-    await jest.isolateModulesAsync(async () => {
-      const { users } = await import('../../src/api/users');
-
-      const router = express();
-
-      router.use((req, res, next) => {
-        req.session = { email: 'pancakes@waffles.bananas' } as any;
-        next();
-      }, users);
-
-      const res = await supertest(router).post('/');
-      expect(res.statusCode).toBe(200);
-    });
+    // setup
+    const req = {
+      ...mockReq ,
+      session: { email: '' }
+    };
+    // test
+    await addUser(req as unknown as Request, mockRes as unknown as Response, mockNext);
+    // assert
+    expect(mockNext).toHaveBeenCalled();
   });
   
   it('returns a 401 when a valid session cannot be found', async () => {
-    await jest.isolateModulesAsync(async () => {
+    await jest.isolateModulesAsync( async () => {
       const { users } = await import('../../src/api/users');
       
       const router = express();
@@ -69,13 +65,12 @@ describe('mounting user on /user', () => {
     };
     // test
     await addUser(req as unknown as Request, mockRes as unknown as Response, mockNext);
-    // await addUser( mockReq as any, mockRes as any, mockNext );
-
+    
     // assert
     expect(req.user).toBeDefined();
     expect(mockNext).toHaveBeenCalled();
-
   });
+  
   it('returns a 403 when a valid session is found but the user is not authorized', async () => {
     // setup
     const req = {
@@ -83,9 +78,10 @@ describe('mounting user on /user', () => {
       session: { email: 'unauthorized@user.io' }
     };
     // test
-    await addUser( mockReq as any, mockRes as any, mockNext );
+    await addUser(req as unknown as Request, mockRes as unknown as Response, mockNext);
     // assert
     expect(mockRes.sendStatus).toHaveBeenCalledWith(403);
     expect(mockNext).not.toHaveBeenCalled();
   });
+
 });
